+++ src/shared/components/toast/Toast.css.ts
@@ -6,7 +6,7 @@ import { colorVars } from '@/shared/styles/tokens/color.css';
 export const container = style({
   display: 'inline-flex',
   gap: '1rem',
-  padding: '10px 20px',
+  padding: '1rem 2rem',
   borderRadius: '30px',
   background: colorVars.color.gray900,
 });
