+++ src/stories/Toast.stories.tsx
@@ -1,6 +1,6 @@
 import { TOAST_TYPE } from '@/shared/types/toast';
 
-import Toast from './Toast';
+import Toast from '../shared/components/toast/Toast';
 
 import type { Meta, StoryObj } from '@storybook/react-vite';
